+++ app/Notifications/CheckEmail.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Notifications;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Notifications\Messages\MailMessage;
+use Illuminate\Notifications\Notification;
+
+class CheckEmail extends Notification
+{
+    use Queueable;
+
+    public function via($notifiable)
+    {
+        return ['mail'];
+    }
+
+    public function toMail($notifiable)
+    {
+        return (new MailMessage)
+                ->subject(__('shaarli.mails.check.title'))
+                ->line(__('shaarli.mails.check.message', ['name' => config('app.name')]));
+    }
+
+    public function toArray($notifiable)
+    {
+        return [];
+    }
+}